Fix 3 bugs caught by first real deployment against live Postgres: V9 rating column type mismatch (SMALLINT vs Integer), and NULL bind-parameter type inference failures in ProjectRepository/UserRepository search queries. Also parameterize docker-compose secrets via .env and move web frontend to port 3001.

+++ agroinvest-backend/src/main/resources/db/migration/V9__reviews.sql
@@ -7,7 +7,7 @@ CREATE TABLE reviews (
     investment_id UUID NOT NULL UNIQUE REFERENCES investments(id),
     investor_id UUID NOT NULL REFERENCES users(id),
     farmer_id UUID NOT NULL REFERENCES users(id),
-    rating SMALLINT NOT NULL CHECK (rating BETWEEN 1 AND 5),
+    rating INTEGER NOT NULL CHECK (rating BETWEEN 1 AND 5),
     comment TEXT,
     created_at TIMESTAMP DEFAULT now()
 );
